More DRY for actions - implement actions and notes and IF base class and use standard form for creating them unless an IF needs special behaviour.
